Refactor bindings to use different file names
Instead of storing  bindings as auto/VERSION/name.ext use
auto/nameVERSION.ext. This avoids issues with moc in some versions of
Qt.
