+++ src/CMakeLists.txt
@@ -1,5 +1,5 @@
 
-set(NEOVIM_QT_SOURCES util.cpp neovimconnector.cpp neovimconnectorhelper.cpp function.cpp msgpackrequest.cpp msgpackiodevice.cpp auto/1/neovim.cpp auto/2/neovim.cpp)
+set(NEOVIM_QT_SOURCES util.cpp neovimconnector.cpp neovimconnectorhelper.cpp function.cpp msgpackrequest.cpp msgpackiodevice.cpp auto/neovimapi1.cpp auto/neovimapi2.cpp)
 if(WIN32)
   list(APPEND NEOVIM_QT_SOURCES stdinreader.cpp)
 endif()
