+++ src/auto/neovimapi1.cpp
@@ -1,5 +1,5 @@
-// Auto generated 2017-10-05 20:54:40.457724 from nvim API level:2
-#include "auto/2/neovim.h"
+// Auto generated 2017-10-06 11:39:25.385122 from nvim API level:1
+#include "auto/neovimapi1.h"
 #include "neovimconnector.h"
 #include "msgpackrequest.h"
 #include "msgpackiodevice.h"
@@ -9,7 +9,7 @@ namespace NeovimQt {
 /* Unpack Neovim EXT types Window, Buffer Tabpage which are all
  * uint64_t see Neovim:msgpack_rpc_to_
  */
-QVariant unpackBufferApi2(MsgpackIODevice *dev, const char* in, quint32 size)
+QVariant unpackBufferApi1(MsgpackIODevice *dev, const char* in, quint32 size)
 {
 	msgpack_unpacked result;
 	msgpack_unpacked_init(&result);
@@ -21,67 +21,67 @@ QVariant unpackBufferApi2(MsgpackIODevice *dev, const char* in, quint32 size)
 	}
 	return QVariant((quint64)result.data.via.u64);
 }
-#define unpackWindowApi2 unpackBufferApi2
-#define unpackTabpageApi2 unpackBufferApi2
+#define unpackWindowApi1 unpackBufferApi1
+#define unpackTabpageApi1 unpackBufferApi1
 
-NeovimApi2::NeovimApi2(NeovimConnector *c)
+NeovimApi1::NeovimApi1(NeovimConnector *c)
 :m_c(c)
 {
 	// EXT types
-		m_c->m_dev->registerExtType(0, unpackBufferApi2);
-		m_c->m_dev->registerExtType(1, unpackWindowApi2);
-		m_c->m_dev->registerExtType(2, unpackTabpageApi2);
+		m_c->m_dev->registerExtType(0, unpackBufferApi1);
+		m_c->m_dev->registerExtType(1, unpackWindowApi1);
+		m_c->m_dev->registerExtType(2, unpackTabpageApi1);
 		connect(m_c->m_dev, &MsgpackIODevice::notification,
-			this, &NeovimApi2::neovimNotification);
+			this, &NeovimApi1::neovimNotification);
 }
 
 // Slots
-MsgpackRequest* NeovimApi2::nvim_buf_line_count(int64_t buffer)
+MsgpackRequest* NeovimApi1::nvim_buf_line_count(int64_t buffer)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_line_count", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_LINE_COUNT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_LINE_COUNT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_get_line(int64_t buffer, int64_t index)
+MsgpackRequest* NeovimApi1::buffer_get_line(int64_t buffer, int64_t index)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_get_line", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_GET_LINE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_GET_LINE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(index);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_set_line(int64_t buffer, int64_t index, QByteArray line)
+MsgpackRequest* NeovimApi1::buffer_set_line(int64_t buffer, int64_t index, QByteArray line)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_set_line", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_SET_LINE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_SET_LINE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(index);
 	m_c->m_dev->send(line);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_del_line(int64_t buffer, int64_t index)
+MsgpackRequest* NeovimApi1::buffer_del_line(int64_t buffer, int64_t index)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_del_line", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_DEL_LINE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_DEL_LINE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(index);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_get_line_slice(int64_t buffer, int64_t start, int64_t end, bool include_start, bool include_end)
+MsgpackRequest* NeovimApi1::buffer_get_line_slice(int64_t buffer, int64_t start, int64_t end, bool include_start, bool include_end)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_get_line_slice", 5);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_GET_LINE_SLICE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_GET_LINE_SLICE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(start);
 	m_c->m_dev->send(end);
@@ -89,24 +89,24 @@ MsgpackRequest* NeovimApi2::buffer_get_line_slice(int64_t buffer, int64_t start,
 	m_c->m_dev->send(include_end);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_get_lines(int64_t buffer, int64_t start, int64_t end, bool strict_indexing)
+MsgpackRequest* NeovimApi1::nvim_buf_get_lines(int64_t buffer, int64_t start, int64_t end, bool strict_indexing)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_get_lines", 4);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_LINES);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_LINES);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(start);
 	m_c->m_dev->send(end);
 	m_c->m_dev->send(strict_indexing);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_set_line_slice(int64_t buffer, int64_t start, int64_t end, bool include_start, bool include_end, QList<QByteArray> replacement)
+MsgpackRequest* NeovimApi1::buffer_set_line_slice(int64_t buffer, int64_t start, int64_t end, bool include_start, bool include_end, QList<QByteArray> replacement)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_set_line_slice", 6);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_SET_LINE_SLICE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_SET_LINE_SLICE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(start);
 	m_c->m_dev->send(end);
@@ -115,12 +115,12 @@ MsgpackRequest* NeovimApi2::buffer_set_line_slice(int64_t buffer, int64_t start,
 	m_c->m_dev->sendArrayOf(replacement);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_set_lines(int64_t buffer, int64_t start, int64_t end, bool strict_indexing, QList<QByteArray> replacement)
+MsgpackRequest* NeovimApi1::nvim_buf_set_lines(int64_t buffer, int64_t start, int64_t end, bool strict_indexing, QList<QByteArray> replacement)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_set_lines", 5);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_SET_LINES);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_SET_LINES);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(start);
 	m_c->m_dev->send(end);
@@ -128,152 +128,143 @@ MsgpackRequest* NeovimApi2::nvim_buf_set_lines(int64_t buffer, int64_t start, in
 	m_c->m_dev->sendArrayOf(replacement);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_get_var(int64_t buffer, QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_buf_get_var(int64_t buffer, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_get_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_get_changedtick(int64_t buffer)
-{
-	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_get_changedtick", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_CHANGEDTICK);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
-	m_c->m_dev->send(buffer);
-	return r;
-}
-MsgpackRequest* NeovimApi2::nvim_buf_set_var(int64_t buffer, QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::nvim_buf_set_var(int64_t buffer, QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_set_var", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_SET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_SET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_del_var(int64_t buffer, QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_buf_del_var(int64_t buffer, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_del_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_DEL_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_DEL_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_set_var(int64_t buffer, QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::buffer_set_var(int64_t buffer, QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_set_var", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_SET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_SET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_del_var(int64_t buffer, QByteArray name)
+MsgpackRequest* NeovimApi1::buffer_del_var(int64_t buffer, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_del_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_DEL_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_DEL_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_get_option(int64_t buffer, QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_buf_get_option(int64_t buffer, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_get_option", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_set_option(int64_t buffer, QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::nvim_buf_set_option(int64_t buffer, QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_set_option", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_SET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_SET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_get_number(int64_t buffer)
+MsgpackRequest* NeovimApi1::nvim_buf_get_number(int64_t buffer)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_get_number", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_NUMBER);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_NUMBER);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_get_name(int64_t buffer)
+MsgpackRequest* NeovimApi1::nvim_buf_get_name(int64_t buffer)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_get_name", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_NAME);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_NAME);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_set_name(int64_t buffer, QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_buf_set_name(int64_t buffer, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_set_name", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_SET_NAME);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_SET_NAME);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_is_valid(int64_t buffer)
+MsgpackRequest* NeovimApi1::nvim_buf_is_valid(int64_t buffer)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_is_valid", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_IS_VALID);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_IS_VALID);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_insert(int64_t buffer, int64_t lnum, QList<QByteArray> lines)
+MsgpackRequest* NeovimApi1::buffer_insert(int64_t buffer, int64_t lnum, QList<QByteArray> lines)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_insert", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_INSERT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_INSERT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(lnum);
 	m_c->m_dev->sendArrayOf(lines);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_get_mark(int64_t buffer, QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_buf_get_mark(int64_t buffer, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_get_mark", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_MARK);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_MARK);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_add_highlight(int64_t buffer, int64_t src_id, QByteArray hl_group, int64_t line, int64_t col_start, int64_t col_end)
+MsgpackRequest* NeovimApi1::nvim_buf_add_highlight(int64_t buffer, int64_t src_id, QByteArray hl_group, int64_t line, int64_t col_start, int64_t col_end)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_add_highlight", 6);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_ADD_HIGHLIGHT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_ADD_HIGHLIGHT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(src_id);
 	m_c->m_dev->send(hl_group);
@@ -282,715 +273,707 @@ MsgpackRequest* NeovimApi2::nvim_buf_add_highlight(int64_t buffer, int64_t src_i
 	m_c->m_dev->send(col_end);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_buf_clear_highlight(int64_t buffer, int64_t src_id, int64_t line_start, int64_t line_end)
+MsgpackRequest* NeovimApi1::nvim_buf_clear_highlight(int64_t buffer, int64_t src_id, int64_t line_start, int64_t line_end)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_buf_clear_highlight", 4);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_BUF_CLEAR_HIGHLIGHT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_BUF_CLEAR_HIGHLIGHT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(src_id);
 	m_c->m_dev->send(line_start);
 	m_c->m_dev->send(line_end);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_tabpage_list_wins(int64_t tabpage)
+MsgpackRequest* NeovimApi1::nvim_tabpage_list_wins(int64_t tabpage)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_tabpage_list_wins", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_LIST_WINS);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_LIST_WINS);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_tabpage_get_var(int64_t tabpage, QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_tabpage_get_var(int64_t tabpage, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_tabpage_get_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_GET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_GET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_tabpage_set_var(int64_t tabpage, QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::nvim_tabpage_set_var(int64_t tabpage, QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_tabpage_set_var", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_SET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_SET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_tabpage_del_var(int64_t tabpage, QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_tabpage_del_var(int64_t tabpage, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_tabpage_del_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_DEL_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_DEL_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::tabpage_set_var(int64_t tabpage, QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::tabpage_set_var(int64_t tabpage, QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("tabpage_set_var", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_TABPAGE_SET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_TABPAGE_SET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::tabpage_del_var(int64_t tabpage, QByteArray name)
+MsgpackRequest* NeovimApi1::tabpage_del_var(int64_t tabpage, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("tabpage_del_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_TABPAGE_DEL_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_TABPAGE_DEL_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_tabpage_get_win(int64_t tabpage)
+MsgpackRequest* NeovimApi1::nvim_tabpage_get_win(int64_t tabpage)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_tabpage_get_win", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_GET_WIN);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_GET_WIN);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_tabpage_get_number(int64_t tabpage)
+MsgpackRequest* NeovimApi1::nvim_tabpage_get_number(int64_t tabpage)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_tabpage_get_number", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_GET_NUMBER);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_GET_NUMBER);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_tabpage_is_valid(int64_t tabpage)
+MsgpackRequest* NeovimApi1::nvim_tabpage_is_valid(int64_t tabpage)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_tabpage_is_valid", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_IS_VALID);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_IS_VALID);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_ui_attach(int64_t width, int64_t height, QVariantMap options)
+MsgpackRequest* NeovimApi1::nvim_ui_attach(int64_t width, int64_t height, QVariantMap options)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_ui_attach", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_UI_ATTACH);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_UI_ATTACH);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(width);
 	m_c->m_dev->send(height);
 	m_c->m_dev->send(options);
 	return r;
 }
-MsgpackRequest* NeovimApi2::ui_attach(int64_t width, int64_t height, bool enable_rgb)
+MsgpackRequest* NeovimApi1::ui_attach(int64_t width, int64_t height, bool enable_rgb)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("ui_attach", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_UI_ATTACH);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_UI_ATTACH);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(width);
 	m_c->m_dev->send(height);
 	m_c->m_dev->send(enable_rgb);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_ui_detach()
+MsgpackRequest* NeovimApi1::nvim_ui_detach()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_ui_detach", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_UI_DETACH);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_UI_DETACH);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_ui_try_resize(int64_t width, int64_t height)
+MsgpackRequest* NeovimApi1::nvim_ui_try_resize(int64_t width, int64_t height)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_ui_try_resize", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_UI_TRY_RESIZE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_UI_TRY_RESIZE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(width);
 	m_c->m_dev->send(height);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_ui_set_option(QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::nvim_ui_set_option(QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_ui_set_option", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_UI_SET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_UI_SET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_command(QByteArray command)
+MsgpackRequest* NeovimApi1::nvim_command(QByteArray command)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_command", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_COMMAND);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_COMMAND);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(command);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_feedkeys(QByteArray keys, QByteArray mode, bool escape_csi)
+MsgpackRequest* NeovimApi1::nvim_feedkeys(QByteArray keys, QByteArray mode, bool escape_csi)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_feedkeys", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_FEEDKEYS);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_FEEDKEYS);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(keys);
 	m_c->m_dev->send(mode);
 	m_c->m_dev->send(escape_csi);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_input(QByteArray keys)
+MsgpackRequest* NeovimApi1::nvim_input(QByteArray keys)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_input", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_INPUT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_INPUT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(keys);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_replace_termcodes(QByteArray str, bool from_part, bool do_lt, bool special)
+MsgpackRequest* NeovimApi1::nvim_replace_termcodes(QByteArray str, bool from_part, bool do_lt, bool special)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_replace_termcodes", 4);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_REPLACE_TERMCODES);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_REPLACE_TERMCODES);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(str);
 	m_c->m_dev->send(from_part);
 	m_c->m_dev->send(do_lt);
 	m_c->m_dev->send(special);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_command_output(QByteArray str)
+MsgpackRequest* NeovimApi1::nvim_command_output(QByteArray str)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_command_output", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_COMMAND_OUTPUT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_COMMAND_OUTPUT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(str);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_eval(QByteArray expr)
+MsgpackRequest* NeovimApi1::nvim_eval(QByteArray expr)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_eval", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_EVAL);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_EVAL);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(expr);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_call_function(QByteArray fname, QVariantList args)
+MsgpackRequest* NeovimApi1::nvim_call_function(QByteArray fname, QVariantList args)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_call_function", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_CALL_FUNCTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_CALL_FUNCTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(fname);
 	m_c->m_dev->send(args);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_strwidth(QByteArray str)
+MsgpackRequest* NeovimApi1::nvim_strwidth(QByteArray str)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_strwidth", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_STRWIDTH);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_STRWIDTH);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(str);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_list_runtime_paths()
+MsgpackRequest* NeovimApi1::nvim_list_runtime_paths()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_list_runtime_paths", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_LIST_RUNTIME_PATHS);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_LIST_RUNTIME_PATHS);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_set_current_dir(QByteArray dir)
+MsgpackRequest* NeovimApi1::nvim_set_current_dir(QByteArray dir)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_set_current_dir", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_DIR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_DIR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(dir);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_get_current_line()
+MsgpackRequest* NeovimApi1::nvim_get_current_line()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_get_current_line", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_GET_CURRENT_LINE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_GET_CURRENT_LINE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_set_current_line(QByteArray line)
+MsgpackRequest* NeovimApi1::nvim_set_current_line(QByteArray line)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_set_current_line", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_LINE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_LINE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(line);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_del_current_line()
+MsgpackRequest* NeovimApi1::nvim_del_current_line()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_del_current_line", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_DEL_CURRENT_LINE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_DEL_CURRENT_LINE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_get_var(QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_get_var(QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_get_var", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_GET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_GET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_set_var(QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::nvim_set_var(QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_set_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_SET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_SET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_del_var(QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_del_var(QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_del_var", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_DEL_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_DEL_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_set_var(QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::vim_set_var(QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_set_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_SET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_SET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_del_var(QByteArray name)
+MsgpackRequest* NeovimApi1::vim_del_var(QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_del_var", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_DEL_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_DEL_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_get_vvar(QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_get_vvar(QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_get_vvar", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_GET_VVAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_GET_VVAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_get_option(QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_get_option(QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_get_option", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_GET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_GET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_set_option(QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::nvim_set_option(QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_set_option", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_SET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_SET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_out_write(QByteArray str)
+MsgpackRequest* NeovimApi1::nvim_out_write(QByteArray str)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_out_write", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_OUT_WRITE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_OUT_WRITE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(str);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_err_write(QByteArray str)
+MsgpackRequest* NeovimApi1::nvim_err_write(QByteArray str)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_err_write", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_ERR_WRITE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_ERR_WRITE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(str);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_err_writeln(QByteArray str)
+MsgpackRequest* NeovimApi1::nvim_err_writeln(QByteArray str)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_err_writeln", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_ERR_WRITELN);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_ERR_WRITELN);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(str);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_list_bufs()
+MsgpackRequest* NeovimApi1::nvim_list_bufs()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_list_bufs", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_LIST_BUFS);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_LIST_BUFS);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_get_current_buf()
+MsgpackRequest* NeovimApi1::nvim_get_current_buf()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_get_current_buf", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_GET_CURRENT_BUF);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_GET_CURRENT_BUF);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_set_current_buf(int64_t buffer)
+MsgpackRequest* NeovimApi1::nvim_set_current_buf(int64_t buffer)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_set_current_buf", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_BUF);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_BUF);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_list_wins()
+MsgpackRequest* NeovimApi1::nvim_list_wins()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_list_wins", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_LIST_WINS);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_LIST_WINS);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_get_current_win()
+MsgpackRequest* NeovimApi1::nvim_get_current_win()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_get_current_win", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_GET_CURRENT_WIN);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_GET_CURRENT_WIN);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_set_current_win(int64_t window)
+MsgpackRequest* NeovimApi1::nvim_set_current_win(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_set_current_win", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_WIN);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_WIN);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_list_tabpages()
+MsgpackRequest* NeovimApi1::nvim_list_tabpages()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_list_tabpages", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_LIST_TABPAGES);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_LIST_TABPAGES);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_get_current_tabpage()
+MsgpackRequest* NeovimApi1::nvim_get_current_tabpage()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_get_current_tabpage", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_GET_CURRENT_TABPAGE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_GET_CURRENT_TABPAGE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_set_current_tabpage(int64_t tabpage)
+MsgpackRequest* NeovimApi1::nvim_set_current_tabpage(int64_t tabpage)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_set_current_tabpage", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_TABPAGE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_TABPAGE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_subscribe(QByteArray event)
+MsgpackRequest* NeovimApi1::nvim_subscribe(QByteArray event)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_subscribe", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_SUBSCRIBE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_SUBSCRIBE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(event);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_unsubscribe(QByteArray event)
+MsgpackRequest* NeovimApi1::nvim_unsubscribe(QByteArray event)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_unsubscribe", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_UNSUBSCRIBE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_UNSUBSCRIBE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(event);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_get_color_by_name(QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_get_color_by_name(QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_get_color_by_name", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_GET_COLOR_BY_NAME);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_GET_COLOR_BY_NAME);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_get_color_map()
+MsgpackRequest* NeovimApi1::nvim_get_color_map()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_get_color_map", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_GET_COLOR_MAP);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_GET_COLOR_MAP);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_get_mode()
-{
-	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_get_mode", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_GET_MODE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
-	return r;
-}
-MsgpackRequest* NeovimApi2::nvim_get_api_info()
+MsgpackRequest* NeovimApi1::nvim_get_api_info()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_get_api_info", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_GET_API_INFO);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_GET_API_INFO);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_call_atomic(QVariantList calls)
+MsgpackRequest* NeovimApi1::nvim_call_atomic(QVariantList calls)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_call_atomic", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_CALL_ATOMIC);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_CALL_ATOMIC);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(calls);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_get_buf(int64_t window)
+MsgpackRequest* NeovimApi1::nvim_win_get_buf(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_get_buf", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_BUF);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_BUF);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_get_cursor(int64_t window)
+MsgpackRequest* NeovimApi1::nvim_win_get_cursor(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_get_cursor", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_CURSOR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_CURSOR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_set_cursor(int64_t window, QPoint pos)
+MsgpackRequest* NeovimApi1::nvim_win_set_cursor(int64_t window, QPoint pos)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_set_cursor", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_CURSOR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_CURSOR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(pos);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_get_height(int64_t window)
+MsgpackRequest* NeovimApi1::nvim_win_get_height(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_get_height", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_HEIGHT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_HEIGHT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_set_height(int64_t window, int64_t height)
+MsgpackRequest* NeovimApi1::nvim_win_set_height(int64_t window, int64_t height)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_set_height", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_HEIGHT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_HEIGHT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(height);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_get_width(int64_t window)
+MsgpackRequest* NeovimApi1::nvim_win_get_width(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_get_width", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_WIDTH);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_WIDTH);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_set_width(int64_t window, int64_t width)
+MsgpackRequest* NeovimApi1::nvim_win_set_width(int64_t window, int64_t width)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_set_width", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_WIDTH);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_WIDTH);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(width);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_get_var(int64_t window, QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_win_get_var(int64_t window, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_get_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_set_var(int64_t window, QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::nvim_win_set_var(int64_t window, QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_set_var", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_del_var(int64_t window, QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_win_del_var(int64_t window, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_del_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_DEL_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_DEL_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_set_var(int64_t window, QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::window_set_var(int64_t window, QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_set_var", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_SET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_SET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_del_var(int64_t window, QByteArray name)
+MsgpackRequest* NeovimApi1::window_del_var(int64_t window, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_del_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_DEL_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_DEL_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_get_option(int64_t window, QByteArray name)
+MsgpackRequest* NeovimApi1::nvim_win_get_option(int64_t window, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_get_option", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_set_option(int64_t window, QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::nvim_win_set_option(int64_t window, QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_set_option", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_get_position(int64_t window)
+MsgpackRequest* NeovimApi1::nvim_win_get_position(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_get_position", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_POSITION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_POSITION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_get_tabpage(int64_t window)
+MsgpackRequest* NeovimApi1::nvim_win_get_tabpage(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_get_tabpage", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_TABPAGE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_TABPAGE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_get_number(int64_t window)
+MsgpackRequest* NeovimApi1::nvim_win_get_number(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_get_number", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_NUMBER);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_NUMBER);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::nvim_win_is_valid(int64_t window)
+MsgpackRequest* NeovimApi1::nvim_win_is_valid(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("nvim_win_is_valid", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_NVIM_WIN_IS_VALID);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_NVIM_WIN_IS_VALID);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_line_count(int64_t buffer)
+MsgpackRequest* NeovimApi1::buffer_line_count(int64_t buffer)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_line_count", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_LINE_COUNT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_LINE_COUNT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_get_lines(int64_t buffer, int64_t start, int64_t end, bool strict_indexing)
+MsgpackRequest* NeovimApi1::buffer_get_lines(int64_t buffer, int64_t start, int64_t end, bool strict_indexing)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_get_lines", 4);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_GET_LINES);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_GET_LINES);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(start);
 	m_c->m_dev->send(end);
 	m_c->m_dev->send(strict_indexing);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_set_lines(int64_t buffer, int64_t start, int64_t end, bool strict_indexing, QList<QByteArray> replacement)
+MsgpackRequest* NeovimApi1::buffer_set_lines(int64_t buffer, int64_t start, int64_t end, bool strict_indexing, QList<QByteArray> replacement)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_set_lines", 5);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_SET_LINES);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_SET_LINES);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(start);
 	m_c->m_dev->send(end);
@@ -998,90 +981,90 @@ MsgpackRequest* NeovimApi2::buffer_set_lines(int64_t buffer, int64_t start, int6
 	m_c->m_dev->sendArrayOf(replacement);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_get_var(int64_t buffer, QByteArray name)
+MsgpackRequest* NeovimApi1::buffer_get_var(int64_t buffer, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_get_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_GET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_GET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_get_option(int64_t buffer, QByteArray name)
+MsgpackRequest* NeovimApi1::buffer_get_option(int64_t buffer, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_get_option", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_GET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_GET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_set_option(int64_t buffer, QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::buffer_set_option(int64_t buffer, QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_set_option", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_SET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_SET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_get_number(int64_t buffer)
+MsgpackRequest* NeovimApi1::buffer_get_number(int64_t buffer)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_get_number", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_GET_NUMBER);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_GET_NUMBER);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_get_name(int64_t buffer)
+MsgpackRequest* NeovimApi1::buffer_get_name(int64_t buffer)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_get_name", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_GET_NAME);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_GET_NAME);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_set_name(int64_t buffer, QByteArray name)
+MsgpackRequest* NeovimApi1::buffer_set_name(int64_t buffer, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_set_name", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_SET_NAME);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_SET_NAME);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_is_valid(int64_t buffer)
+MsgpackRequest* NeovimApi1::buffer_is_valid(int64_t buffer)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_is_valid", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_IS_VALID);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_IS_VALID);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_get_mark(int64_t buffer, QByteArray name)
+MsgpackRequest* NeovimApi1::buffer_get_mark(int64_t buffer, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_get_mark", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_GET_MARK);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_GET_MARK);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_add_highlight(int64_t buffer, int64_t src_id, QByteArray hl_group, int64_t line, int64_t col_start, int64_t col_end)
+MsgpackRequest* NeovimApi1::buffer_add_highlight(int64_t buffer, int64_t src_id, QByteArray hl_group, int64_t line, int64_t col_start, int64_t col_end)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_add_highlight", 6);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_ADD_HIGHLIGHT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_ADD_HIGHLIGHT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(src_id);
 	m_c->m_dev->send(hl_group);
@@ -1090,495 +1073,495 @@ MsgpackRequest* NeovimApi2::buffer_add_highlight(int64_t buffer, int64_t src_id,
 	m_c->m_dev->send(col_end);
 	return r;
 }
-MsgpackRequest* NeovimApi2::buffer_clear_highlight(int64_t buffer, int64_t src_id, int64_t line_start, int64_t line_end)
+MsgpackRequest* NeovimApi1::buffer_clear_highlight(int64_t buffer, int64_t src_id, int64_t line_start, int64_t line_end)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("buffer_clear_highlight", 4);
-	r->setFunction(NeovimApi2::NEOVIM_FN_BUFFER_CLEAR_HIGHLIGHT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_BUFFER_CLEAR_HIGHLIGHT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	m_c->m_dev->send(src_id);
 	m_c->m_dev->send(line_start);
 	m_c->m_dev->send(line_end);
 	return r;
 }
-MsgpackRequest* NeovimApi2::tabpage_get_windows(int64_t tabpage)
+MsgpackRequest* NeovimApi1::tabpage_get_windows(int64_t tabpage)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("tabpage_get_windows", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_TABPAGE_GET_WINDOWS);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_TABPAGE_GET_WINDOWS);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	return r;
 }
-MsgpackRequest* NeovimApi2::tabpage_get_var(int64_t tabpage, QByteArray name)
+MsgpackRequest* NeovimApi1::tabpage_get_var(int64_t tabpage, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("tabpage_get_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_TABPAGE_GET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_TABPAGE_GET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::tabpage_get_window(int64_t tabpage)
+MsgpackRequest* NeovimApi1::tabpage_get_window(int64_t tabpage)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("tabpage_get_window", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_TABPAGE_GET_WINDOW);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_TABPAGE_GET_WINDOW);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	return r;
 }
-MsgpackRequest* NeovimApi2::tabpage_is_valid(int64_t tabpage)
+MsgpackRequest* NeovimApi1::tabpage_is_valid(int64_t tabpage)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("tabpage_is_valid", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_TABPAGE_IS_VALID);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_TABPAGE_IS_VALID);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	return r;
 }
-MsgpackRequest* NeovimApi2::ui_detach()
+MsgpackRequest* NeovimApi1::ui_detach()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("ui_detach", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_UI_DETACH);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_UI_DETACH);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::ui_try_resize(int64_t width, int64_t height)
+MsgpackRequest* NeovimApi1::ui_try_resize(int64_t width, int64_t height)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("ui_try_resize", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_UI_TRY_RESIZE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_UI_TRY_RESIZE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(width);
 	m_c->m_dev->send(height);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_command(QByteArray command)
+MsgpackRequest* NeovimApi1::vim_command(QByteArray command)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_command", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_COMMAND);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_COMMAND);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(command);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_feedkeys(QByteArray keys, QByteArray mode, bool escape_csi)
+MsgpackRequest* NeovimApi1::vim_feedkeys(QByteArray keys, QByteArray mode, bool escape_csi)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_feedkeys", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_FEEDKEYS);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_FEEDKEYS);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(keys);
 	m_c->m_dev->send(mode);
 	m_c->m_dev->send(escape_csi);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_input(QByteArray keys)
+MsgpackRequest* NeovimApi1::vim_input(QByteArray keys)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_input", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_INPUT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_INPUT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(keys);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_replace_termcodes(QByteArray str, bool from_part, bool do_lt, bool special)
+MsgpackRequest* NeovimApi1::vim_replace_termcodes(QByteArray str, bool from_part, bool do_lt, bool special)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_replace_termcodes", 4);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_REPLACE_TERMCODES);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_REPLACE_TERMCODES);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(str);
 	m_c->m_dev->send(from_part);
 	m_c->m_dev->send(do_lt);
 	m_c->m_dev->send(special);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_command_output(QByteArray str)
+MsgpackRequest* NeovimApi1::vim_command_output(QByteArray str)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_command_output", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_COMMAND_OUTPUT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_COMMAND_OUTPUT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(str);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_eval(QByteArray expr)
+MsgpackRequest* NeovimApi1::vim_eval(QByteArray expr)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_eval", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_EVAL);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_EVAL);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(expr);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_call_function(QByteArray fname, QVariantList args)
+MsgpackRequest* NeovimApi1::vim_call_function(QByteArray fname, QVariantList args)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_call_function", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_CALL_FUNCTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_CALL_FUNCTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(fname);
 	m_c->m_dev->send(args);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_strwidth(QByteArray str)
+MsgpackRequest* NeovimApi1::vim_strwidth(QByteArray str)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_strwidth", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_STRWIDTH);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_STRWIDTH);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(str);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_list_runtime_paths()
+MsgpackRequest* NeovimApi1::vim_list_runtime_paths()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_list_runtime_paths", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_LIST_RUNTIME_PATHS);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_LIST_RUNTIME_PATHS);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_change_directory(QByteArray dir)
+MsgpackRequest* NeovimApi1::vim_change_directory(QByteArray dir)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_change_directory", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_CHANGE_DIRECTORY);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_CHANGE_DIRECTORY);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(dir);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_get_current_line()
+MsgpackRequest* NeovimApi1::vim_get_current_line()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_get_current_line", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_GET_CURRENT_LINE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_GET_CURRENT_LINE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_set_current_line(QByteArray line)
+MsgpackRequest* NeovimApi1::vim_set_current_line(QByteArray line)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_set_current_line", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_SET_CURRENT_LINE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_SET_CURRENT_LINE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(line);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_del_current_line()
+MsgpackRequest* NeovimApi1::vim_del_current_line()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_del_current_line", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_DEL_CURRENT_LINE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_DEL_CURRENT_LINE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_get_var(QByteArray name)
+MsgpackRequest* NeovimApi1::vim_get_var(QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_get_var", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_GET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_GET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_get_vvar(QByteArray name)
+MsgpackRequest* NeovimApi1::vim_get_vvar(QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_get_vvar", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_GET_VVAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_GET_VVAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_get_option(QByteArray name)
+MsgpackRequest* NeovimApi1::vim_get_option(QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_get_option", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_GET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_GET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_set_option(QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::vim_set_option(QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_set_option", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_SET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_SET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_out_write(QByteArray str)
+MsgpackRequest* NeovimApi1::vim_out_write(QByteArray str)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_out_write", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_OUT_WRITE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_OUT_WRITE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(str);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_err_write(QByteArray str)
+MsgpackRequest* NeovimApi1::vim_err_write(QByteArray str)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_err_write", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_ERR_WRITE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_ERR_WRITE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(str);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_report_error(QByteArray str)
+MsgpackRequest* NeovimApi1::vim_report_error(QByteArray str)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_report_error", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_REPORT_ERROR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_REPORT_ERROR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(str);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_get_buffers()
+MsgpackRequest* NeovimApi1::vim_get_buffers()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_get_buffers", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_GET_BUFFERS);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_GET_BUFFERS);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_get_current_buffer()
+MsgpackRequest* NeovimApi1::vim_get_current_buffer()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_get_current_buffer", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_GET_CURRENT_BUFFER);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_GET_CURRENT_BUFFER);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_set_current_buffer(int64_t buffer)
+MsgpackRequest* NeovimApi1::vim_set_current_buffer(int64_t buffer)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_set_current_buffer", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_SET_CURRENT_BUFFER);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_SET_CURRENT_BUFFER);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(buffer);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_get_windows()
+MsgpackRequest* NeovimApi1::vim_get_windows()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_get_windows", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_GET_WINDOWS);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_GET_WINDOWS);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_get_current_window()
+MsgpackRequest* NeovimApi1::vim_get_current_window()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_get_current_window", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_GET_CURRENT_WINDOW);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_GET_CURRENT_WINDOW);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_set_current_window(int64_t window)
+MsgpackRequest* NeovimApi1::vim_set_current_window(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_set_current_window", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_SET_CURRENT_WINDOW);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_SET_CURRENT_WINDOW);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_get_tabpages()
+MsgpackRequest* NeovimApi1::vim_get_tabpages()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_get_tabpages", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_GET_TABPAGES);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_GET_TABPAGES);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_get_current_tabpage()
+MsgpackRequest* NeovimApi1::vim_get_current_tabpage()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_get_current_tabpage", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_GET_CURRENT_TABPAGE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_GET_CURRENT_TABPAGE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_set_current_tabpage(int64_t tabpage)
+MsgpackRequest* NeovimApi1::vim_set_current_tabpage(int64_t tabpage)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_set_current_tabpage", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_SET_CURRENT_TABPAGE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_SET_CURRENT_TABPAGE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(tabpage);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_subscribe(QByteArray event)
+MsgpackRequest* NeovimApi1::vim_subscribe(QByteArray event)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_subscribe", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_SUBSCRIBE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_SUBSCRIBE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(event);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_unsubscribe(QByteArray event)
+MsgpackRequest* NeovimApi1::vim_unsubscribe(QByteArray event)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_unsubscribe", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_UNSUBSCRIBE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_UNSUBSCRIBE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(event);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_name_to_color(QByteArray name)
+MsgpackRequest* NeovimApi1::vim_name_to_color(QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_name_to_color", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_NAME_TO_COLOR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_NAME_TO_COLOR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::vim_get_color_map()
+MsgpackRequest* NeovimApi1::vim_get_color_map()
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("vim_get_color_map", 0);
-	r->setFunction(NeovimApi2::NEOVIM_FN_VIM_GET_COLOR_MAP);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_VIM_GET_COLOR_MAP);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_get_buffer(int64_t window)
+MsgpackRequest* NeovimApi1::window_get_buffer(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_get_buffer", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_GET_BUFFER);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_GET_BUFFER);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_get_cursor(int64_t window)
+MsgpackRequest* NeovimApi1::window_get_cursor(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_get_cursor", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_GET_CURSOR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_GET_CURSOR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_set_cursor(int64_t window, QPoint pos)
+MsgpackRequest* NeovimApi1::window_set_cursor(int64_t window, QPoint pos)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_set_cursor", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_SET_CURSOR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_SET_CURSOR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(pos);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_get_height(int64_t window)
+MsgpackRequest* NeovimApi1::window_get_height(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_get_height", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_GET_HEIGHT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_GET_HEIGHT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_set_height(int64_t window, int64_t height)
+MsgpackRequest* NeovimApi1::window_set_height(int64_t window, int64_t height)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_set_height", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_SET_HEIGHT);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_SET_HEIGHT);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(height);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_get_width(int64_t window)
+MsgpackRequest* NeovimApi1::window_get_width(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_get_width", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_GET_WIDTH);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_GET_WIDTH);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_set_width(int64_t window, int64_t width)
+MsgpackRequest* NeovimApi1::window_set_width(int64_t window, int64_t width)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_set_width", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_SET_WIDTH);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_SET_WIDTH);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(width);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_get_var(int64_t window, QByteArray name)
+MsgpackRequest* NeovimApi1::window_get_var(int64_t window, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_get_var", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_GET_VAR);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_GET_VAR);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_get_option(int64_t window, QByteArray name)
+MsgpackRequest* NeovimApi1::window_get_option(int64_t window, QByteArray name)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_get_option", 2);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_GET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_GET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(name);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_set_option(int64_t window, QByteArray name, QVariant value)
+MsgpackRequest* NeovimApi1::window_set_option(int64_t window, QByteArray name, QVariant value)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_set_option", 3);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_SET_OPTION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_SET_OPTION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	m_c->m_dev->send(name);
 	m_c->m_dev->send(value);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_get_position(int64_t window)
+MsgpackRequest* NeovimApi1::window_get_position(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_get_position", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_GET_POSITION);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_GET_POSITION);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_get_tabpage(int64_t window)
+MsgpackRequest* NeovimApi1::window_get_tabpage(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_get_tabpage", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_GET_TABPAGE);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_GET_TABPAGE);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
-MsgpackRequest* NeovimApi2::window_is_valid(int64_t window)
+MsgpackRequest* NeovimApi1::window_is_valid(int64_t window)
 {
 	MsgpackRequest *r = m_c->m_dev->startRequestUnchecked("window_is_valid", 1);
-	r->setFunction(NeovimApi2::NEOVIM_FN_WINDOW_IS_VALID);
-	connect(r, &MsgpackRequest::finished, this, &NeovimApi2::handleResponse);
-	connect(r, &MsgpackRequest::error, this, &NeovimApi2::handleResponseError);
+	r->setFunction(NeovimApi1::NEOVIM_FN_WINDOW_IS_VALID);
+	connect(r, &MsgpackRequest::finished, this, &NeovimApi1::handleResponse);
+	connect(r, &MsgpackRequest::error, this, &NeovimApi1::handleResponseError);
 	m_c->m_dev->send(window);
 	return r;
 }
 
 // Handlers
 
-void NeovimApi2::handleResponseError(quint32 msgid, quint64 fun, const QVariant& res)
+void NeovimApi1::handleResponseError(quint32 msgid, quint64 fun, const QVariant& res)
 {
 
 	// TODO: support Neovim error types Exception/Validation/etc
@@ -1593,487 +1576,481 @@ void NeovimApi2::handleResponseError(quint32 msgid, quint64 fun, const QVariant&
 	}
 
 	switch(fun) {
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_LINE_COUNT:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_LINE_COUNT:
 		emit err_nvim_buf_line_count(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_LINE:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_LINE:
 		emit err_buffer_get_line(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_SET_LINE:
+	case NeovimApi1::NEOVIM_FN_BUFFER_SET_LINE:
 		emit err_buffer_set_line(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_DEL_LINE:
+	case NeovimApi1::NEOVIM_FN_BUFFER_DEL_LINE:
 		emit err_buffer_del_line(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_LINE_SLICE:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_LINE_SLICE:
 		emit err_buffer_get_line_slice(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_LINES:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_LINES:
 		emit err_nvim_buf_get_lines(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_SET_LINE_SLICE:
+	case NeovimApi1::NEOVIM_FN_BUFFER_SET_LINE_SLICE:
 		emit err_buffer_set_line_slice(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_SET_LINES:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_SET_LINES:
 		emit err_nvim_buf_set_lines(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_VAR:
 		emit err_nvim_buf_get_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_CHANGEDTICK:
-		emit err_nvim_buf_get_changedtick(errMsg, res);
-		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_SET_VAR:
 		emit err_nvim_buf_set_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_DEL_VAR:
 		emit err_nvim_buf_del_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_BUFFER_SET_VAR:
 		emit err_buffer_set_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_BUFFER_DEL_VAR:
 		emit err_buffer_del_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_OPTION:
 		emit err_nvim_buf_get_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_SET_OPTION:
 		emit err_nvim_buf_set_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_NUMBER:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_NUMBER:
 		emit err_nvim_buf_get_number(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_NAME:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_NAME:
 		emit err_nvim_buf_get_name(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_SET_NAME:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_SET_NAME:
 		emit err_nvim_buf_set_name(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_IS_VALID:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_IS_VALID:
 		emit err_nvim_buf_is_valid(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_INSERT:
+	case NeovimApi1::NEOVIM_FN_BUFFER_INSERT:
 		emit err_buffer_insert(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_MARK:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_MARK:
 		emit err_nvim_buf_get_mark(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_ADD_HIGHLIGHT:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_ADD_HIGHLIGHT:
 		emit err_nvim_buf_add_highlight(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_CLEAR_HIGHLIGHT:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_CLEAR_HIGHLIGHT:
 		emit err_nvim_buf_clear_highlight(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_LIST_WINS:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_LIST_WINS:
 		emit err_nvim_tabpage_list_wins(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_GET_VAR:
 		emit err_nvim_tabpage_get_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_SET_VAR:
 		emit err_nvim_tabpage_set_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_DEL_VAR:
 		emit err_nvim_tabpage_del_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_TABPAGE_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_TABPAGE_SET_VAR:
 		emit err_tabpage_set_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_TABPAGE_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_TABPAGE_DEL_VAR:
 		emit err_tabpage_del_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_GET_WIN:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_GET_WIN:
 		emit err_nvim_tabpage_get_win(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_GET_NUMBER:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_GET_NUMBER:
 		emit err_nvim_tabpage_get_number(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_IS_VALID:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_IS_VALID:
 		emit err_nvim_tabpage_is_valid(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_UI_ATTACH:
+	case NeovimApi1::NEOVIM_FN_NVIM_UI_ATTACH:
 		emit err_nvim_ui_attach(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_UI_ATTACH:
+	case NeovimApi1::NEOVIM_FN_UI_ATTACH:
 		emit err_ui_attach(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_UI_DETACH:
+	case NeovimApi1::NEOVIM_FN_NVIM_UI_DETACH:
 		emit err_nvim_ui_detach(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_UI_TRY_RESIZE:
+	case NeovimApi1::NEOVIM_FN_NVIM_UI_TRY_RESIZE:
 		emit err_nvim_ui_try_resize(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_UI_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_UI_SET_OPTION:
 		emit err_nvim_ui_set_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_COMMAND:
+	case NeovimApi1::NEOVIM_FN_NVIM_COMMAND:
 		emit err_nvim_command(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_FEEDKEYS:
+	case NeovimApi1::NEOVIM_FN_NVIM_FEEDKEYS:
 		emit err_nvim_feedkeys(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_INPUT:
+	case NeovimApi1::NEOVIM_FN_NVIM_INPUT:
 		emit err_nvim_input(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_REPLACE_TERMCODES:
+	case NeovimApi1::NEOVIM_FN_NVIM_REPLACE_TERMCODES:
 		emit err_nvim_replace_termcodes(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_COMMAND_OUTPUT:
+	case NeovimApi1::NEOVIM_FN_NVIM_COMMAND_OUTPUT:
 		emit err_nvim_command_output(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_EVAL:
+	case NeovimApi1::NEOVIM_FN_NVIM_EVAL:
 		emit err_nvim_eval(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_CALL_FUNCTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_CALL_FUNCTION:
 		emit err_nvim_call_function(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_STRWIDTH:
+	case NeovimApi1::NEOVIM_FN_NVIM_STRWIDTH:
 		emit err_nvim_strwidth(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_LIST_RUNTIME_PATHS:
+	case NeovimApi1::NEOVIM_FN_NVIM_LIST_RUNTIME_PATHS:
 		emit err_nvim_list_runtime_paths(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_DIR:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_DIR:
 		emit err_nvim_set_current_dir(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_CURRENT_LINE:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_CURRENT_LINE:
 		emit err_nvim_get_current_line(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_LINE:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_LINE:
 		emit err_nvim_set_current_line(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_DEL_CURRENT_LINE:
+	case NeovimApi1::NEOVIM_FN_NVIM_DEL_CURRENT_LINE:
 		emit err_nvim_del_current_line(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_VAR:
 		emit err_nvim_get_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_VAR:
 		emit err_nvim_set_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_DEL_VAR:
 		emit err_nvim_del_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_VIM_SET_VAR:
 		emit err_vim_set_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_VIM_DEL_VAR:
 		emit err_vim_del_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_VVAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_VVAR:
 		emit err_nvim_get_vvar(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_OPTION:
 		emit err_nvim_get_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_OPTION:
 		emit err_nvim_set_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_OUT_WRITE:
+	case NeovimApi1::NEOVIM_FN_NVIM_OUT_WRITE:
 		emit err_nvim_out_write(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_ERR_WRITE:
+	case NeovimApi1::NEOVIM_FN_NVIM_ERR_WRITE:
 		emit err_nvim_err_write(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_ERR_WRITELN:
+	case NeovimApi1::NEOVIM_FN_NVIM_ERR_WRITELN:
 		emit err_nvim_err_writeln(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_LIST_BUFS:
+	case NeovimApi1::NEOVIM_FN_NVIM_LIST_BUFS:
 		emit err_nvim_list_bufs(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_CURRENT_BUF:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_CURRENT_BUF:
 		emit err_nvim_get_current_buf(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_BUF:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_BUF:
 		emit err_nvim_set_current_buf(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_LIST_WINS:
+	case NeovimApi1::NEOVIM_FN_NVIM_LIST_WINS:
 		emit err_nvim_list_wins(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_CURRENT_WIN:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_CURRENT_WIN:
 		emit err_nvim_get_current_win(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_WIN:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_WIN:
 		emit err_nvim_set_current_win(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_LIST_TABPAGES:
+	case NeovimApi1::NEOVIM_FN_NVIM_LIST_TABPAGES:
 		emit err_nvim_list_tabpages(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_CURRENT_TABPAGE:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_CURRENT_TABPAGE:
 		emit err_nvim_get_current_tabpage(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_TABPAGE:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_TABPAGE:
 		emit err_nvim_set_current_tabpage(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SUBSCRIBE:
+	case NeovimApi1::NEOVIM_FN_NVIM_SUBSCRIBE:
 		emit err_nvim_subscribe(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_UNSUBSCRIBE:
+	case NeovimApi1::NEOVIM_FN_NVIM_UNSUBSCRIBE:
 		emit err_nvim_unsubscribe(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_COLOR_BY_NAME:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_COLOR_BY_NAME:
 		emit err_nvim_get_color_by_name(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_COLOR_MAP:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_COLOR_MAP:
 		emit err_nvim_get_color_map(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_MODE:
-		emit err_nvim_get_mode(errMsg, res);
-		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_API_INFO:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_API_INFO:
 		emit err_nvim_get_api_info(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_CALL_ATOMIC:
+	case NeovimApi1::NEOVIM_FN_NVIM_CALL_ATOMIC:
 		emit err_nvim_call_atomic(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_BUF:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_BUF:
 		emit err_nvim_win_get_buf(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_CURSOR:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_CURSOR:
 		emit err_nvim_win_get_cursor(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_CURSOR:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_CURSOR:
 		emit err_nvim_win_set_cursor(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_HEIGHT:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_HEIGHT:
 		emit err_nvim_win_get_height(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_HEIGHT:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_HEIGHT:
 		emit err_nvim_win_set_height(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_WIDTH:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_WIDTH:
 		emit err_nvim_win_get_width(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_WIDTH:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_WIDTH:
 		emit err_nvim_win_set_width(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_VAR:
 		emit err_nvim_win_get_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_VAR:
 		emit err_nvim_win_set_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_DEL_VAR:
 		emit err_nvim_win_del_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_WINDOW_SET_VAR:
 		emit err_window_set_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_WINDOW_DEL_VAR:
 		emit err_window_del_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_OPTION:
 		emit err_nvim_win_get_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_OPTION:
 		emit err_nvim_win_set_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_POSITION:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_POSITION:
 		emit err_nvim_win_get_position(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_TABPAGE:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_TABPAGE:
 		emit err_nvim_win_get_tabpage(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_NUMBER:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_NUMBER:
 		emit err_nvim_win_get_number(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_IS_VALID:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_IS_VALID:
 		emit err_nvim_win_is_valid(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_LINE_COUNT:
+	case NeovimApi1::NEOVIM_FN_BUFFER_LINE_COUNT:
 		emit err_buffer_line_count(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_LINES:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_LINES:
 		emit err_buffer_get_lines(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_SET_LINES:
+	case NeovimApi1::NEOVIM_FN_BUFFER_SET_LINES:
 		emit err_buffer_set_lines(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_VAR:
 		emit err_buffer_get_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_OPTION:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_OPTION:
 		emit err_buffer_get_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_BUFFER_SET_OPTION:
 		emit err_buffer_set_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_NUMBER:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_NUMBER:
 		emit err_buffer_get_number(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_NAME:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_NAME:
 		emit err_buffer_get_name(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_SET_NAME:
+	case NeovimApi1::NEOVIM_FN_BUFFER_SET_NAME:
 		emit err_buffer_set_name(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_IS_VALID:
+	case NeovimApi1::NEOVIM_FN_BUFFER_IS_VALID:
 		emit err_buffer_is_valid(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_MARK:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_MARK:
 		emit err_buffer_get_mark(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_ADD_HIGHLIGHT:
+	case NeovimApi1::NEOVIM_FN_BUFFER_ADD_HIGHLIGHT:
 		emit err_buffer_add_highlight(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_CLEAR_HIGHLIGHT:
+	case NeovimApi1::NEOVIM_FN_BUFFER_CLEAR_HIGHLIGHT:
 		emit err_buffer_clear_highlight(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_TABPAGE_GET_WINDOWS:
+	case NeovimApi1::NEOVIM_FN_TABPAGE_GET_WINDOWS:
 		emit err_tabpage_get_windows(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_TABPAGE_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_TABPAGE_GET_VAR:
 		emit err_tabpage_get_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_TABPAGE_GET_WINDOW:
+	case NeovimApi1::NEOVIM_FN_TABPAGE_GET_WINDOW:
 		emit err_tabpage_get_window(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_TABPAGE_IS_VALID:
+	case NeovimApi1::NEOVIM_FN_TABPAGE_IS_VALID:
 		emit err_tabpage_is_valid(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_UI_DETACH:
+	case NeovimApi1::NEOVIM_FN_UI_DETACH:
 		emit err_ui_detach(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_UI_TRY_RESIZE:
+	case NeovimApi1::NEOVIM_FN_UI_TRY_RESIZE:
 		emit err_ui_try_resize(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_COMMAND:
+	case NeovimApi1::NEOVIM_FN_VIM_COMMAND:
 		emit err_vim_command(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_FEEDKEYS:
+	case NeovimApi1::NEOVIM_FN_VIM_FEEDKEYS:
 		emit err_vim_feedkeys(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_INPUT:
+	case NeovimApi1::NEOVIM_FN_VIM_INPUT:
 		emit err_vim_input(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_REPLACE_TERMCODES:
+	case NeovimApi1::NEOVIM_FN_VIM_REPLACE_TERMCODES:
 		emit err_vim_replace_termcodes(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_COMMAND_OUTPUT:
+	case NeovimApi1::NEOVIM_FN_VIM_COMMAND_OUTPUT:
 		emit err_vim_command_output(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_EVAL:
+	case NeovimApi1::NEOVIM_FN_VIM_EVAL:
 		emit err_vim_eval(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_CALL_FUNCTION:
+	case NeovimApi1::NEOVIM_FN_VIM_CALL_FUNCTION:
 		emit err_vim_call_function(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_STRWIDTH:
+	case NeovimApi1::NEOVIM_FN_VIM_STRWIDTH:
 		emit err_vim_strwidth(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_LIST_RUNTIME_PATHS:
+	case NeovimApi1::NEOVIM_FN_VIM_LIST_RUNTIME_PATHS:
 		emit err_vim_list_runtime_paths(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_CHANGE_DIRECTORY:
+	case NeovimApi1::NEOVIM_FN_VIM_CHANGE_DIRECTORY:
 		emit err_vim_change_directory(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_CURRENT_LINE:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_CURRENT_LINE:
 		emit err_vim_get_current_line(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SET_CURRENT_LINE:
+	case NeovimApi1::NEOVIM_FN_VIM_SET_CURRENT_LINE:
 		emit err_vim_set_current_line(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_DEL_CURRENT_LINE:
+	case NeovimApi1::NEOVIM_FN_VIM_DEL_CURRENT_LINE:
 		emit err_vim_del_current_line(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_VAR:
 		emit err_vim_get_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_VVAR:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_VVAR:
 		emit err_vim_get_vvar(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_OPTION:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_OPTION:
 		emit err_vim_get_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_VIM_SET_OPTION:
 		emit err_vim_set_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_OUT_WRITE:
+	case NeovimApi1::NEOVIM_FN_VIM_OUT_WRITE:
 		emit err_vim_out_write(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_ERR_WRITE:
+	case NeovimApi1::NEOVIM_FN_VIM_ERR_WRITE:
 		emit err_vim_err_write(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_REPORT_ERROR:
+	case NeovimApi1::NEOVIM_FN_VIM_REPORT_ERROR:
 		emit err_vim_report_error(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_BUFFERS:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_BUFFERS:
 		emit err_vim_get_buffers(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_CURRENT_BUFFER:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_CURRENT_BUFFER:
 		emit err_vim_get_current_buffer(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SET_CURRENT_BUFFER:
+	case NeovimApi1::NEOVIM_FN_VIM_SET_CURRENT_BUFFER:
 		emit err_vim_set_current_buffer(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_WINDOWS:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_WINDOWS:
 		emit err_vim_get_windows(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_CURRENT_WINDOW:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_CURRENT_WINDOW:
 		emit err_vim_get_current_window(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SET_CURRENT_WINDOW:
+	case NeovimApi1::NEOVIM_FN_VIM_SET_CURRENT_WINDOW:
 		emit err_vim_set_current_window(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_TABPAGES:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_TABPAGES:
 		emit err_vim_get_tabpages(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_CURRENT_TABPAGE:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_CURRENT_TABPAGE:
 		emit err_vim_get_current_tabpage(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SET_CURRENT_TABPAGE:
+	case NeovimApi1::NEOVIM_FN_VIM_SET_CURRENT_TABPAGE:
 		emit err_vim_set_current_tabpage(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SUBSCRIBE:
+	case NeovimApi1::NEOVIM_FN_VIM_SUBSCRIBE:
 		emit err_vim_subscribe(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_UNSUBSCRIBE:
+	case NeovimApi1::NEOVIM_FN_VIM_UNSUBSCRIBE:
 		emit err_vim_unsubscribe(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_NAME_TO_COLOR:
+	case NeovimApi1::NEOVIM_FN_VIM_NAME_TO_COLOR:
 		emit err_vim_name_to_color(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_COLOR_MAP:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_COLOR_MAP:
 		emit err_vim_get_color_map(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_BUFFER:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_BUFFER:
 		emit err_window_get_buffer(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_CURSOR:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_CURSOR:
 		emit err_window_get_cursor(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_SET_CURSOR:
+	case NeovimApi1::NEOVIM_FN_WINDOW_SET_CURSOR:
 		emit err_window_set_cursor(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_HEIGHT:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_HEIGHT:
 		emit err_window_get_height(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_SET_HEIGHT:
+	case NeovimApi1::NEOVIM_FN_WINDOW_SET_HEIGHT:
 		emit err_window_set_height(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_WIDTH:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_WIDTH:
 		emit err_window_get_width(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_SET_WIDTH:
+	case NeovimApi1::NEOVIM_FN_WINDOW_SET_WIDTH:
 		emit err_window_set_width(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_VAR:
 		emit err_window_get_var(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_OPTION:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_OPTION:
 		emit err_window_get_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_WINDOW_SET_OPTION:
 		emit err_window_set_option(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_POSITION:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_POSITION:
 		emit err_window_get_position(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_TABPAGE:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_TABPAGE:
 		emit err_window_get_tabpage(errMsg, res);
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_IS_VALID:
+	case NeovimApi1::NEOVIM_FN_WINDOW_IS_VALID:
 		emit err_window_is_valid(errMsg, res);
 		break;
 	default:
@@ -2081,10 +2058,10 @@ void NeovimApi2::handleResponseError(quint32 msgid, quint64 fun, const QVariant&
 	}
 }
 
-void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
+void NeovimApi1::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 {
 	switch(fun) {
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_LINE_COUNT:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_LINE_COUNT:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2096,7 +2073,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_LINE:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_LINE:
 		{
 			QByteArray data;
 			if (decode(res, data)) {
@@ -2108,19 +2085,19 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_SET_LINE:
+	case NeovimApi1::NEOVIM_FN_BUFFER_SET_LINE:
 		{
 			emit on_buffer_set_line();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_DEL_LINE:
+	case NeovimApi1::NEOVIM_FN_BUFFER_DEL_LINE:
 		{
 			emit on_buffer_del_line();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_LINE_SLICE:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_LINE_SLICE:
 		{
 			QList<QByteArray> data;
 			if (decode(res, data)) {
@@ -2132,7 +2109,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_LINES:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_LINES:
 		{
 			QList<QByteArray> data;
 			if (decode(res, data)) {
@@ -2144,19 +2121,19 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_SET_LINE_SLICE:
+	case NeovimApi1::NEOVIM_FN_BUFFER_SET_LINE_SLICE:
 		{
 			emit on_buffer_set_line_slice();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_SET_LINES:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_SET_LINES:
 		{
 			emit on_nvim_buf_set_lines();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2168,31 +2145,19 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_CHANGEDTICK:
-		{
-			int64_t data;
-			if (decode(res, data)) {
-				m_c->setError(NeovimConnector::RuntimeMsgpackError, "Error unpacking return type for nvim_buf_get_changedtick");
-				return;
-			} else {
-				emit on_nvim_buf_get_changedtick(data);
-			}
-
-		}
-		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_SET_VAR:
 		{
 			emit on_nvim_buf_set_var();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_DEL_VAR:
 		{
 			emit on_nvim_buf_del_var();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_BUFFER_SET_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2204,7 +2169,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_BUFFER_DEL_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2216,7 +2181,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_OPTION:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2228,13 +2193,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_SET_OPTION:
 		{
 			emit on_nvim_buf_set_option();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_NUMBER:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_NUMBER:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2246,7 +2211,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_NAME:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_NAME:
 		{
 			QByteArray data;
 			if (decode(res, data)) {
@@ -2258,13 +2223,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_SET_NAME:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_SET_NAME:
 		{
 			emit on_nvim_buf_set_name();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_IS_VALID:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_IS_VALID:
 		{
 			bool data;
 			if (decode(res, data)) {
@@ -2276,13 +2241,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_INSERT:
+	case NeovimApi1::NEOVIM_FN_BUFFER_INSERT:
 		{
 			emit on_buffer_insert();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_GET_MARK:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_GET_MARK:
 		{
 			QPoint data;
 			if (decode(res, data)) {
@@ -2294,7 +2259,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_ADD_HIGHLIGHT:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_ADD_HIGHLIGHT:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2306,13 +2271,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_BUF_CLEAR_HIGHLIGHT:
+	case NeovimApi1::NEOVIM_FN_NVIM_BUF_CLEAR_HIGHLIGHT:
 		{
 			emit on_nvim_buf_clear_highlight();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_LIST_WINS:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_LIST_WINS:
 		{
 			QList<int64_t> data;
 			if (decode(res, data)) {
@@ -2324,7 +2289,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_GET_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2336,19 +2301,19 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_SET_VAR:
 		{
 			emit on_nvim_tabpage_set_var();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_DEL_VAR:
 		{
 			emit on_nvim_tabpage_del_var();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_TABPAGE_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_TABPAGE_SET_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2360,7 +2325,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_TABPAGE_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_TABPAGE_DEL_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2372,7 +2337,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_GET_WIN:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_GET_WIN:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2384,7 +2349,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_GET_NUMBER:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_GET_NUMBER:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2396,7 +2361,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_TABPAGE_IS_VALID:
+	case NeovimApi1::NEOVIM_FN_NVIM_TABPAGE_IS_VALID:
 		{
 			bool data;
 			if (decode(res, data)) {
@@ -2408,49 +2373,49 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_UI_ATTACH:
+	case NeovimApi1::NEOVIM_FN_NVIM_UI_ATTACH:
 		{
 			emit on_nvim_ui_attach();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_UI_ATTACH:
+	case NeovimApi1::NEOVIM_FN_UI_ATTACH:
 		{
 			emit on_ui_attach();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_UI_DETACH:
+	case NeovimApi1::NEOVIM_FN_NVIM_UI_DETACH:
 		{
 			emit on_nvim_ui_detach();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_UI_TRY_RESIZE:
+	case NeovimApi1::NEOVIM_FN_NVIM_UI_TRY_RESIZE:
 		{
 			emit on_nvim_ui_try_resize();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_UI_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_UI_SET_OPTION:
 		{
 			emit on_nvim_ui_set_option();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_COMMAND:
+	case NeovimApi1::NEOVIM_FN_NVIM_COMMAND:
 		{
 			emit on_nvim_command();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_FEEDKEYS:
+	case NeovimApi1::NEOVIM_FN_NVIM_FEEDKEYS:
 		{
 			emit on_nvim_feedkeys();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_INPUT:
+	case NeovimApi1::NEOVIM_FN_NVIM_INPUT:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2462,7 +2427,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_REPLACE_TERMCODES:
+	case NeovimApi1::NEOVIM_FN_NVIM_REPLACE_TERMCODES:
 		{
 			QByteArray data;
 			if (decode(res, data)) {
@@ -2474,7 +2439,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_COMMAND_OUTPUT:
+	case NeovimApi1::NEOVIM_FN_NVIM_COMMAND_OUTPUT:
 		{
 			QByteArray data;
 			if (decode(res, data)) {
@@ -2486,7 +2451,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_EVAL:
+	case NeovimApi1::NEOVIM_FN_NVIM_EVAL:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2498,7 +2463,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_CALL_FUNCTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_CALL_FUNCTION:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2510,7 +2475,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_STRWIDTH:
+	case NeovimApi1::NEOVIM_FN_NVIM_STRWIDTH:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2522,7 +2487,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_LIST_RUNTIME_PATHS:
+	case NeovimApi1::NEOVIM_FN_NVIM_LIST_RUNTIME_PATHS:
 		{
 			QList<QByteArray> data;
 			if (decode(res, data)) {
@@ -2534,13 +2499,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_DIR:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_DIR:
 		{
 			emit on_nvim_set_current_dir();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_CURRENT_LINE:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_CURRENT_LINE:
 		{
 			QByteArray data;
 			if (decode(res, data)) {
@@ -2552,19 +2517,19 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_LINE:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_LINE:
 		{
 			emit on_nvim_set_current_line();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_DEL_CURRENT_LINE:
+	case NeovimApi1::NEOVIM_FN_NVIM_DEL_CURRENT_LINE:
 		{
 			emit on_nvim_del_current_line();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2576,19 +2541,19 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_VAR:
 		{
 			emit on_nvim_set_var();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_DEL_VAR:
 		{
 			emit on_nvim_del_var();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_VIM_SET_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2600,7 +2565,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_VIM_DEL_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2612,7 +2577,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_VVAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_VVAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2624,7 +2589,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_OPTION:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2636,31 +2601,31 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_OPTION:
 		{
 			emit on_nvim_set_option();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_OUT_WRITE:
+	case NeovimApi1::NEOVIM_FN_NVIM_OUT_WRITE:
 		{
 			emit on_nvim_out_write();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_ERR_WRITE:
+	case NeovimApi1::NEOVIM_FN_NVIM_ERR_WRITE:
 		{
 			emit on_nvim_err_write();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_ERR_WRITELN:
+	case NeovimApi1::NEOVIM_FN_NVIM_ERR_WRITELN:
 		{
 			emit on_nvim_err_writeln();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_LIST_BUFS:
+	case NeovimApi1::NEOVIM_FN_NVIM_LIST_BUFS:
 		{
 			QList<int64_t> data;
 			if (decode(res, data)) {
@@ -2672,7 +2637,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_CURRENT_BUF:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_CURRENT_BUF:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2684,13 +2649,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_BUF:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_BUF:
 		{
 			emit on_nvim_set_current_buf();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_LIST_WINS:
+	case NeovimApi1::NEOVIM_FN_NVIM_LIST_WINS:
 		{
 			QList<int64_t> data;
 			if (decode(res, data)) {
@@ -2702,7 +2667,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_CURRENT_WIN:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_CURRENT_WIN:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2714,13 +2679,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_WIN:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_WIN:
 		{
 			emit on_nvim_set_current_win();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_LIST_TABPAGES:
+	case NeovimApi1::NEOVIM_FN_NVIM_LIST_TABPAGES:
 		{
 			QList<int64_t> data;
 			if (decode(res, data)) {
@@ -2732,7 +2697,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_CURRENT_TABPAGE:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_CURRENT_TABPAGE:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2744,25 +2709,25 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SET_CURRENT_TABPAGE:
+	case NeovimApi1::NEOVIM_FN_NVIM_SET_CURRENT_TABPAGE:
 		{
 			emit on_nvim_set_current_tabpage();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_SUBSCRIBE:
+	case NeovimApi1::NEOVIM_FN_NVIM_SUBSCRIBE:
 		{
 			emit on_nvim_subscribe();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_UNSUBSCRIBE:
+	case NeovimApi1::NEOVIM_FN_NVIM_UNSUBSCRIBE:
 		{
 			emit on_nvim_unsubscribe();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_COLOR_BY_NAME:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_COLOR_BY_NAME:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2774,7 +2739,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_COLOR_MAP:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_COLOR_MAP:
 		{
 			QVariantMap data;
 			if (decode(res, data)) {
@@ -2786,19 +2751,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_MODE:
-		{
-			QVariantMap data;
-			if (decode(res, data)) {
-				m_c->setError(NeovimConnector::RuntimeMsgpackError, "Error unpacking return type for nvim_get_mode");
-				return;
-			} else {
-				emit on_nvim_get_mode(data);
-			}
-
-		}
-		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_GET_API_INFO:
+	case NeovimApi1::NEOVIM_FN_NVIM_GET_API_INFO:
 		{
 			QVariantList data;
 			if (decode(res, data)) {
@@ -2810,7 +2763,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_CALL_ATOMIC:
+	case NeovimApi1::NEOVIM_FN_NVIM_CALL_ATOMIC:
 		{
 			QVariantList data;
 			if (decode(res, data)) {
@@ -2822,7 +2775,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_BUF:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_BUF:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2834,7 +2787,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_CURSOR:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_CURSOR:
 		{
 			QPoint data;
 			if (decode(res, data)) {
@@ -2846,13 +2799,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_CURSOR:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_CURSOR:
 		{
 			emit on_nvim_win_set_cursor();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_HEIGHT:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_HEIGHT:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2864,13 +2817,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_HEIGHT:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_HEIGHT:
 		{
 			emit on_nvim_win_set_height();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_WIDTH:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_WIDTH:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2882,13 +2835,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_WIDTH:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_WIDTH:
 		{
 			emit on_nvim_win_set_width();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2900,19 +2853,19 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_VAR:
 		{
 			emit on_nvim_win_set_var();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_DEL_VAR:
 		{
 			emit on_nvim_win_del_var();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_SET_VAR:
+	case NeovimApi1::NEOVIM_FN_WINDOW_SET_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2924,7 +2877,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_DEL_VAR:
+	case NeovimApi1::NEOVIM_FN_WINDOW_DEL_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2936,7 +2889,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_OPTION:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -2948,13 +2901,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_SET_OPTION:
 		{
 			emit on_nvim_win_set_option();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_POSITION:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_POSITION:
 		{
 			QPoint data;
 			if (decode(res, data)) {
@@ -2966,7 +2919,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_TABPAGE:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_TABPAGE:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2978,7 +2931,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_GET_NUMBER:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_GET_NUMBER:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -2990,7 +2943,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_NVIM_WIN_IS_VALID:
+	case NeovimApi1::NEOVIM_FN_NVIM_WIN_IS_VALID:
 		{
 			bool data;
 			if (decode(res, data)) {
@@ -3002,7 +2955,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_LINE_COUNT:
+	case NeovimApi1::NEOVIM_FN_BUFFER_LINE_COUNT:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3014,7 +2967,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_LINES:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_LINES:
 		{
 			QList<QByteArray> data;
 			if (decode(res, data)) {
@@ -3026,13 +2979,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_SET_LINES:
+	case NeovimApi1::NEOVIM_FN_BUFFER_SET_LINES:
 		{
 			emit on_buffer_set_lines();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -3044,7 +2997,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_OPTION:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_OPTION:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -3056,13 +3009,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_BUFFER_SET_OPTION:
 		{
 			emit on_buffer_set_option();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_NUMBER:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_NUMBER:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3074,7 +3027,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_NAME:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_NAME:
 		{
 			QByteArray data;
 			if (decode(res, data)) {
@@ -3086,13 +3039,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_SET_NAME:
+	case NeovimApi1::NEOVIM_FN_BUFFER_SET_NAME:
 		{
 			emit on_buffer_set_name();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_IS_VALID:
+	case NeovimApi1::NEOVIM_FN_BUFFER_IS_VALID:
 		{
 			bool data;
 			if (decode(res, data)) {
@@ -3104,7 +3057,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_GET_MARK:
+	case NeovimApi1::NEOVIM_FN_BUFFER_GET_MARK:
 		{
 			QPoint data;
 			if (decode(res, data)) {
@@ -3116,7 +3069,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_ADD_HIGHLIGHT:
+	case NeovimApi1::NEOVIM_FN_BUFFER_ADD_HIGHLIGHT:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3128,13 +3081,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_BUFFER_CLEAR_HIGHLIGHT:
+	case NeovimApi1::NEOVIM_FN_BUFFER_CLEAR_HIGHLIGHT:
 		{
 			emit on_buffer_clear_highlight();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_TABPAGE_GET_WINDOWS:
+	case NeovimApi1::NEOVIM_FN_TABPAGE_GET_WINDOWS:
 		{
 			QList<int64_t> data;
 			if (decode(res, data)) {
@@ -3146,7 +3099,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_TABPAGE_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_TABPAGE_GET_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -3158,7 +3111,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_TABPAGE_GET_WINDOW:
+	case NeovimApi1::NEOVIM_FN_TABPAGE_GET_WINDOW:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3170,7 +3123,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_TABPAGE_IS_VALID:
+	case NeovimApi1::NEOVIM_FN_TABPAGE_IS_VALID:
 		{
 			bool data;
 			if (decode(res, data)) {
@@ -3182,13 +3135,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_UI_DETACH:
+	case NeovimApi1::NEOVIM_FN_UI_DETACH:
 		{
 			emit on_ui_detach();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_UI_TRY_RESIZE:
+	case NeovimApi1::NEOVIM_FN_UI_TRY_RESIZE:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -3200,19 +3153,19 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_COMMAND:
+	case NeovimApi1::NEOVIM_FN_VIM_COMMAND:
 		{
 			emit on_vim_command();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_FEEDKEYS:
+	case NeovimApi1::NEOVIM_FN_VIM_FEEDKEYS:
 		{
 			emit on_vim_feedkeys();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_INPUT:
+	case NeovimApi1::NEOVIM_FN_VIM_INPUT:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3224,7 +3177,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_REPLACE_TERMCODES:
+	case NeovimApi1::NEOVIM_FN_VIM_REPLACE_TERMCODES:
 		{
 			QByteArray data;
 			if (decode(res, data)) {
@@ -3236,7 +3189,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_COMMAND_OUTPUT:
+	case NeovimApi1::NEOVIM_FN_VIM_COMMAND_OUTPUT:
 		{
 			QByteArray data;
 			if (decode(res, data)) {
@@ -3248,7 +3201,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_EVAL:
+	case NeovimApi1::NEOVIM_FN_VIM_EVAL:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -3260,7 +3213,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_CALL_FUNCTION:
+	case NeovimApi1::NEOVIM_FN_VIM_CALL_FUNCTION:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -3272,7 +3225,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_STRWIDTH:
+	case NeovimApi1::NEOVIM_FN_VIM_STRWIDTH:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3284,7 +3237,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_LIST_RUNTIME_PATHS:
+	case NeovimApi1::NEOVIM_FN_VIM_LIST_RUNTIME_PATHS:
 		{
 			QList<QByteArray> data;
 			if (decode(res, data)) {
@@ -3296,13 +3249,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_CHANGE_DIRECTORY:
+	case NeovimApi1::NEOVIM_FN_VIM_CHANGE_DIRECTORY:
 		{
 			emit on_vim_change_directory();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_CURRENT_LINE:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_CURRENT_LINE:
 		{
 			QByteArray data;
 			if (decode(res, data)) {
@@ -3314,19 +3267,19 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SET_CURRENT_LINE:
+	case NeovimApi1::NEOVIM_FN_VIM_SET_CURRENT_LINE:
 		{
 			emit on_vim_set_current_line();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_DEL_CURRENT_LINE:
+	case NeovimApi1::NEOVIM_FN_VIM_DEL_CURRENT_LINE:
 		{
 			emit on_vim_del_current_line();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -3338,7 +3291,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_VVAR:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_VVAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -3350,7 +3303,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_OPTION:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_OPTION:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -3362,31 +3315,31 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_VIM_SET_OPTION:
 		{
 			emit on_vim_set_option();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_OUT_WRITE:
+	case NeovimApi1::NEOVIM_FN_VIM_OUT_WRITE:
 		{
 			emit on_vim_out_write();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_ERR_WRITE:
+	case NeovimApi1::NEOVIM_FN_VIM_ERR_WRITE:
 		{
 			emit on_vim_err_write();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_REPORT_ERROR:
+	case NeovimApi1::NEOVIM_FN_VIM_REPORT_ERROR:
 		{
 			emit on_vim_report_error();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_BUFFERS:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_BUFFERS:
 		{
 			QList<int64_t> data;
 			if (decode(res, data)) {
@@ -3398,7 +3351,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_CURRENT_BUFFER:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_CURRENT_BUFFER:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3410,13 +3363,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SET_CURRENT_BUFFER:
+	case NeovimApi1::NEOVIM_FN_VIM_SET_CURRENT_BUFFER:
 		{
 			emit on_vim_set_current_buffer();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_WINDOWS:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_WINDOWS:
 		{
 			QList<int64_t> data;
 			if (decode(res, data)) {
@@ -3428,7 +3381,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_CURRENT_WINDOW:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_CURRENT_WINDOW:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3440,13 +3393,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SET_CURRENT_WINDOW:
+	case NeovimApi1::NEOVIM_FN_VIM_SET_CURRENT_WINDOW:
 		{
 			emit on_vim_set_current_window();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_TABPAGES:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_TABPAGES:
 		{
 			QList<int64_t> data;
 			if (decode(res, data)) {
@@ -3458,7 +3411,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_CURRENT_TABPAGE:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_CURRENT_TABPAGE:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3470,25 +3423,25 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SET_CURRENT_TABPAGE:
+	case NeovimApi1::NEOVIM_FN_VIM_SET_CURRENT_TABPAGE:
 		{
 			emit on_vim_set_current_tabpage();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_SUBSCRIBE:
+	case NeovimApi1::NEOVIM_FN_VIM_SUBSCRIBE:
 		{
 			emit on_vim_subscribe();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_UNSUBSCRIBE:
+	case NeovimApi1::NEOVIM_FN_VIM_UNSUBSCRIBE:
 		{
 			emit on_vim_unsubscribe();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_NAME_TO_COLOR:
+	case NeovimApi1::NEOVIM_FN_VIM_NAME_TO_COLOR:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3500,7 +3453,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_VIM_GET_COLOR_MAP:
+	case NeovimApi1::NEOVIM_FN_VIM_GET_COLOR_MAP:
 		{
 			QVariantMap data;
 			if (decode(res, data)) {
@@ -3512,7 +3465,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_BUFFER:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_BUFFER:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3524,7 +3477,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_CURSOR:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_CURSOR:
 		{
 			QPoint data;
 			if (decode(res, data)) {
@@ -3536,13 +3489,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_SET_CURSOR:
+	case NeovimApi1::NEOVIM_FN_WINDOW_SET_CURSOR:
 		{
 			emit on_window_set_cursor();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_HEIGHT:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_HEIGHT:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3554,13 +3507,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_SET_HEIGHT:
+	case NeovimApi1::NEOVIM_FN_WINDOW_SET_HEIGHT:
 		{
 			emit on_window_set_height();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_WIDTH:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_WIDTH:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3572,13 +3525,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_SET_WIDTH:
+	case NeovimApi1::NEOVIM_FN_WINDOW_SET_WIDTH:
 		{
 			emit on_window_set_width();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_VAR:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_VAR:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -3590,7 +3543,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_OPTION:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_OPTION:
 		{
 			QVariant data;
 			if (decode(res, data)) {
@@ -3602,13 +3555,13 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_SET_OPTION:
+	case NeovimApi1::NEOVIM_FN_WINDOW_SET_OPTION:
 		{
 			emit on_window_set_option();
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_POSITION:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_POSITION:
 		{
 			QPoint data;
 			if (decode(res, data)) {
@@ -3620,7 +3573,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_GET_TABPAGE:
+	case NeovimApi1::NEOVIM_FN_WINDOW_GET_TABPAGE:
 		{
 			int64_t data;
 			if (decode(res, data)) {
@@ -3632,7 +3585,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
 
 		}
 		break;
-	case NeovimApi2::NEOVIM_FN_WINDOW_IS_VALID:
+	case NeovimApi1::NEOVIM_FN_WINDOW_IS_VALID:
 		{
 			bool data;
 			if (decode(res, data)) {
@@ -3657,7 +3610,7 @@ void NeovimApi2::handleResponse(quint32 msgid, quint64 fun, const QVariant& res)
  *
  * Returns false if there is an API mismatch
  */
-bool NeovimApi2::checkFunctions(const QVariantList& ftable)
+bool NeovimApi1::checkFunctions(const QVariantList& ftable)
 {
 
 	QList<Function> required;
@@ -3718,10 +3671,6 @@ bool NeovimApi2::checkFunctions(const QVariantList& ftable)
 			QList<QString>()
 						<< QString("Buffer")
 						<< QString("String")
-						, false)
-		<< Function("Integer", "nvim_buf_get_changedtick",
-			QList<QString>()
-						<< QString("Buffer")
 						, false)
 		<< Function("void", "nvim_buf_set_var",
 			QList<QString>()
@@ -4013,9 +3962,6 @@ bool NeovimApi2::checkFunctions(const QVariantList& ftable)
 						<< QString("String")
 						, false)
 		<< Function("Dictionary", "nvim_get_color_map",
-			QList<QString>()
-						, false)
-		<< Function("Dictionary", "nvim_get_mode",
 			QList<QString>()
 						, false)
 		<< Function("Array", "nvim_get_api_info",
@@ -4409,14 +4355,14 @@ bool NeovimApi2::checkFunctions(const QVariantList& ftable)
 		supported.append(f);
 
 		if (!required.contains(f)) {
-			qDebug() << "Unknown function(api 2)" << f;
+			qDebug() << "Unknown function(api 1)" << f;
 		}
 	}
 
 	bool ok = true;
 	foreach(const Function& f, required) {
 		if (!supported.contains(f)) {
-			qDebug() << "- instance DOES NOT support API2:" << f;
+			qDebug() << "- instance DOES NOT support API1:" << f;
 			ok = false;
 		}
 	}
